requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
httpx
pydantic-settings
sqlalchemy
asyncpg
pyyaml